Code cleanup from pull request and merge of nosql tables for phonenumber and provided income

+++ src/main/java/com/group7/banking/dto/SignUpDTO.java
@@ -11,7 +11,7 @@ public class SignUpDTO {
 	private String lastName;
 	private String birthDate;
 	private String ssn;
-	private Float providedIncome;
+	private Double providedIncome;
 	private String phoneNumber;
 	private String emailAddress;
 	private String addressLine1;
